Ejercicio con dtos, refaztorización con repositorio y crud genérico hecho.

+++ src/main/java/com/softtek/streamshibernate230523/model/Product.java
@@ -31,7 +31,7 @@ public class Product {
     private String quantityPerUnit;
 
     @Column(name = "unit_price")
-    private Float unitPrice;
+    private Double unitPrice;
 
     @Column(name = "units_in_stock")
     private Short unitsInStock;
